pyproject.toml
[project]
name = "easypayments-site"
version = "0.1.0"
description = ""
authors = [
    {name = "dev-ib-byte",email = "dev-ib@kusaka.email"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.125.0,<0.126.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "dependency-injector (>=4.48.3,<5.0.0)",
    "uptrace (>=1.39.0,<2.0.0)",
    "opentelemetry-sdk (>=1.39.1,<2.0.0)",
    "opentelemetry-api (>=1.39.1,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.60b1,<0.61)",
    "opentelemetry-exporter-otlp (>=1.39.1,<2.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "greenlet (>=3.3.0,<4.0.0)",
    "opentelemetry-proto (>=1.39.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "ruff (>=0.14.9,<0.15.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pre-commit-hooks (>=6.0.0,<7.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]

[tool.ruff]
exclude = [
    ".venv",
    "venv",
    "*/__init__.py",
    "*/schema/*",
    "*/tests/*",
    "*/router.py",
    "*/schemas.py",
    "*/migrations/*",
    "src/translations/*",
    "*/replen_optimizer/tasks.py",
    "*/general/management/*",
    "*/exceptions/*",
    "*/calculators/*",
    ".gitignore",
    "*.env*",
    "*/.gitlab*",
    "*/Dockerfile",
    "*/alembic*",
    "*/docker-*",
    "*/README*",
    "*/CHANGELOG.md",
    "app/alembic/*",
    "*/migrations/"
]

line-length = 108
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "D"]

ignore = [
    "D100", "D101", "D102", "D103", "D104", "D105",
    "D106", "D107", "D200", "D204", "D205", "D400",
    "E731", "E722", "F821", "E402", "F405", "F403",
    "D208", "E203",
]


fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F405"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.pycodestyle]
max-line-length = 108

[tool.black]
line-length = 108
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 108

known_first_party = ["config", "infrastructure", "domain", "application", "api"]
known_third_party = ["fastapi", "sqlalchemy", "pydantic", "dependency_injector"]

# Порядок секций (без дубликатов)
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

skip_glob = ["*/migrations/*", "*/alembic/*"]

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
no_implicit_optional = true
explicit_package_bases=true
disable_error_code = ["misc"]
exclude = [
  ".venv",
  "venv",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-vv -ra --cov=src --cov-branch --cov-report=term --cov-report=html --cov-report=xml"
log_cli = false
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:Unmanaged access of declarative attribute.*:sqlalchemy.exc.SAWarning",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.coverage.run]
source = ["src"]
omit = [
    "**/__init__.py",
    "src/main.py",
    "src/worker.py",
    "**/base.py",
    "**/interfaces/*.py",
    "**/fake/*.py",
    "**/fake.py",
    "src/config/*",
    "**/db.py",
    "**/schemas.py",
    "**/models/**",
    "**/exceptions.py",
    "**/dto.py",
    "**/enums.py",
    "**/factories/**",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "__init__",
    "from .* import .*",
    "router = .*",
]
